+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Added the ability to filter via `agent_id` in the filter of a GraphQL query
   requesting a sysmon events.
+- Supports the `PEFile` protocol.
 
 ### Changed